Monitors implemented
- Classes for using in handler from the Tiktok
- Separed legacy files
- New purpose for main page
