--- chat.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Crazy Lab - TikTok Chat</title>
-    <link rel="stylesheet" href="../general.css">
-    <link rel="stylesheet" href="./chat.css">
-    <script defer src="https://unpkg.com/@streamerbot/client/dist/streamerbot-client.js"></script>
-    <script defer src="template.js"></script>
-</head>
-<body>
-    <div id="chatLog"></div>
-</body>
-</html>
