+++ js/modules/tiktok/goalsMonitor.js
@@ -1,0 +1,12 @@
+export class GoalsMonitor {
+  constructor() {
+    this.likes = 0;
+    this.followers = 0;
+    // Etc...
+  }
+
+  updateLikes(newTotalLikes) {
+    this.likes = newTotalLikes;
+    // TODO: Update UI and other features
+  }
+}
