+++ olds_files/goals.html
@@ -1,0 +1,40 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Crazy Lab - TikTok Goals</title>
+    <link rel="stylesheet" href="../general.css">
+    <link rel="stylesheet" href="./styles.css">
+    <script defer src="https://unpkg.com/@streamerbot/client/dist/streamerbot-client.js"></script>
+    <script defer src="goals.js"></script>
+</head>
+<body>
+    <div id="goalFollow" class="goalBox">
+        <div id="goalInformation">
+            <span id="goalFollowTitle">Meta de Seguidores </span>
+            <span id="goalFollowCurrent">0</span>/
+            <span id="goalFollowTarget">0</span>
+        </div>
+
+        <div id="goalFollowBar">
+            <div id="goalFollowPercent">0%</div>
+            <div id="goalFollowBarFill"></div>
+        </div>
+    </div>
+
+    <div id="goalGift" class="goalBox">
+        <div id="goalInformation">
+            <span id="goalGiftTitle">Meta de Gifts </span>
+            <span id="goalGiftCurrent">0</span>/
+            <span id="goalGiftTarget">0</span>
+        </div>
+
+        <div id="goalGiftBar">
+            <div id="goalGiftPercent">0%</div>
+            <div id="goalGiftBarFill"></div>
+        </div>
+    </div>
+</body>
+</html>
